webapp/backend/requirements.txt
# Security-First RAG Web Application Dependencies
# All packages vetted for local-only operation with no external dependencies

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Vector database support
pgvector==0.2.4

# Authentication and security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.1.2  # Python 3.12 compatible with passlib

# Async support
aiofiles==23.2.0
aiohttp==3.9.1

# Data processing (from existing requirements)
numpy>=1.26.0  # Python 3.12 compatible
scikit-learn==1.3.0
pandas==2.0.3

# Text processing and ML (existing)
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# HTTP requests for Ollama communication
requests==2.31.0
httpx==0.25.2

# Configuration management
python-dotenv==1.0.0
pydantic[email]==2.5.0

# File handling
python-magic==0.4.27
filetype==1.2.0

# Logging and monitoring
structlog==23.2.0

# Development and testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Security scanning
bandit==1.7.5
safety==2.3.4

# Performance monitoring
memory-profiler==0.61.0
psutil==5.9.6

# Additional utilities
click==8.1.7
rich==13.7.0
typer==0.9.0

# WebSocket support
websockets==12.0
python-socketio==5.10.0

# Local embedding models (sentence-transformers)
sentence-transformers==2.2.2
torch>=2.2.0  # Python 3.12 compatible
transformers==4.35.2

# Optional: For PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Optional: For document processing
python-docx==1.1.0
openpyxl==3.1.2

# Date and time utilities
python-dateutil==2.8.2
pytz==2023.3

# Validation and serialization
marshmallow==3.20.1
cerberus==1.3.5

# Rate limiting
slowapi==0.1.9

# CORS and middleware
starlette==0.27.0

# Environment detection
python-decouple==3.8